Timestampable trait added
