+++ src/Entity/Traits/TimestampableTrait.php
@@ -1,0 +1,66 @@
+<?php
+
+
+namespace App\Entity\Traits;
+
+use DateTime;
+
+trait TimestampableTrait
+{
+    /**
+     * @ORM\Column(type="datetime")
+     */
+    private $createdAt;
+    /**
+     * @ORM\Column(type="datetime")
+     */
+    private $updatedAt;
+    /**
+     * @return \DateTimeInterface|null
+     */
+    public function getCreatedAt(): ?\DateTimeInterface
+    {
+        return $this->createdAt;
+    }
+
+    /**
+     * @param \DateTimeInterface $createdAt
+     * @return $this
+     */
+    public function setCreatedAt(\DateTimeInterface $createdAt): self
+    {
+        $this->createdAt = $createdAt;
+
+        return $this;
+    }
+
+    /**
+     * @return \DateTimeInterface|null
+     */
+    public function getUpdatedAt(): ?\DateTimeInterface
+    {
+        return $this->updatedAt;
+    }
+
+    /**
+     * @param \DateTimeInterface $updatedAt
+     * @return $this
+     */
+    public function setUpdatedAt(\DateTimeInterface $updatedAt): self
+    {
+        $this->updatedAt = $updatedAt;
+
+        return $this;
+    }
+    /**
+     * @ORM\PrePersist()
+     * @ORM\PreUpdate()
+     */
+    public function updateTimestamps()
+    {
+        $this->updatedAt = new \DateTime('now');
+        if ($this->getCreatedAt() === null) {
+            $this->createdAt = new \DateTime('now');
+        }
+    }
+}
